+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- Added a `LogLevel` property to the client logger. Set to the desired log level to filter the client output logs.
+
+### Changed
+- If you have implemented a custom `ILogger` for Nakama, you must now implement the required `LogLevel` field of the `ILogger` interface.
+
 ## [3.0.0] - 2021-07-14
 ### Added
 - The language tag for the user can be configured with the socket on connect.